fix: メッセージを投稿する場所を独立した画面に変更
Fixes #496

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/rainbeek/isekai/issues/496?shareId=XXXX-XXXX-XXXX-XXXX).
